[Chore] refactoring setups
