+++ Api/run.py
@@ -1,3 +1,4 @@
+import os
 from app import create_app
 
 config_name = "development"
